Redirect www subdomain
www.muncieevents.com and muncieevents.com have separate sessions, so a
user logging into one would find that he's not logged in to the other.
While this move will cause problems with implementing a cookie-free
subdomain for serving static content in the future, it seems to be a
good move now.